refactor: remove procedure executor from DDL manager
